Moved functions into separte pages and then added imports to use on index.js. Added to gitignore.
